+++ Modelo/TipoDeAccion.cs
@@ -17,7 +17,7 @@ namespace PagoAgilFrba.Modelo
         void setearTituloVentana(AbstractForm form);
         void agregarItem(AltaModifFactura form);
         void borrarSeleccionados(AltaModifFactura form);
-        int calcularMontoTotalFactura(AltaModifFactura form);
+        double calcularMontoTotalFactura(AltaModifFactura form);
         List<ItemFactura> armarListaDeItems(AltaModifFactura form);
         void mostrarMensajeDNI(AltaModifCliente form);
     }
